ibm/requirements_noise.txt
qiskit
matplotlib
pylatexenc
numpy
python-dotenv
ipykernel
qiskit_aer
qiskit-ibm-runtime